Fix packet writing field issue.
Adding field to packet caused assert if the field was too large to fit and it failed to allocate space due to a null pointer.  The null pointer was then referenced when trying to rollback the insertion but failed because there was nothing to rollback.  Validating the pointer in the rollback case.